PROJETO-FINAL_CURSO/requirements.txt
Django==5.1.3
python-decouple==3.8
django-crispy-forms==2.3
crispy-bootstrap5==2024.10
django-localflavor==4.0
